Add ability to add/remove playlist / items
Although some methods to handle playlists were already present,
this commit makes them more coherent by adding separate modules
to deal with whole playlists (create, find, delete) and to deal
with playlist items (add, remove).

The prototype of find_or_create_playlists changes, and the yield
is dropped because it was confusing and not really used.