Removed concept of "scope" from authentication
Although it sounded like a good idea, scopes were only used to
raise an error in case the refresh token of the test user had not
been granted all the scopes required to run the tests.

In general, if someone invokes Yt::Account with existing credentials,
we have to assume that those credentials match the right scope.

The only place left for scopes is when it's up to Yt to redirect the
user to the OAuth page. In that case, scopes still make sense because
they have to be passed to the OAuth page, for the user to approve.