Better organization of Error classes
One "public" Yt::Error class is left, which corresponds to
Yt::Errors::RequestError, which is raised when a request to
YouTube does not respond as expected.

Yt::Errors::NoItems is raised when the response is successful
but does not contain items and was supposed to (for instance, when
getting the id or the detail set of a resource).

Yt::Errors::MissingAuth is raised when YouTube responds with failed
authentication.

The code previously tried to avoid making a request to YouTube at all
if the authentication was missing, but Annotations do not require
authentication, so that behavior has been changed.