New ContentOwner with access to partnered channels
Now it’s possible to access the channels managed by a content owner.

ContentOwner is a subclass of Account (because that’s what content
owners are on YouTube) which a #partnered_channels method to list
the partnered channels.

Note that #earnings method on channels now requires to be authorized
as a ContentOwner instance, rather than as a generic Account instance.