Don’t raise error if subscribing to your channel
YouTube does not let people subscribe to their own channels and, even worse,
raises a 500 error in this case, rather than a 4xx.

Instead of bubbling this error up, this commit changes the behavior of
`channel.subscribe` to simply return nil in the case described above.

`channel.subscribe!`, on the other hand, will still raise the error.