Allow angle brackets when creating playlists
Angle brackets are not characters allowed by YouTube, but instead
of failing, Yt replaces them with Unicode characters that look very
similar and are accepted by YouTube.

While the previous commit focused on the "update" action, this commit
focuses on the "insert" action. For now, Yt only allows insertion of
playlists and playlist items, and the latter do not have a snippet, so
the code only changes for playlists, but the code has been restructured
so it will allow to do the same when new #insert methods are added.