Temporarily ignore request specs in Travis
For some undocumented reason, YouTube API is currently returning
500 error for request specs to subscribe. Since some pull requests
unrelated to requests are waiting for tests to pass on Travis CI,
those tests are temporarily skipped to allow those PRs to be accepted.
This will cause code coverage to go down, but it's a temporary fix
waiting for YouTube API to work again.