`size` doesn’t load all the pages of collection
Calling `size` on a collection (for instance on account.videos to know the
number of videos owned by an account) will now return the number of items as
specified in the "totalResults" field of the first page of the YouTube
response, rather than loading *all* the pages (possibly thousands) and counting
exactly how many items are returned.

The old behavior is kept the same for the `count` method.