Remove encrypted keys from .travis.yml
When tests fail, the log file prints out the failing request in
curl format, which might contain authorization tokens that should
not be made public.

Until we have a logger level that allows us not to print out secure
information in error logs, we disable Travis on pull requests and
we run the tests manually.