Add `asset.ownership` to list owners of an asset
This commit adds `has_one :ownership` to the Asset model, which allows to
retrieve the Ownership of an asset.

Ownership includes different types of right owners (general, performance,
synchronization, mechanical), all of which are exposed to the asset by calling

    asset.general_owners
    asset.performance_owners
    asset.synchronization_owners
    asset.mechanical_owners

Each RightOwner includes the owner name, the ratio of ownership and the list
of included/excluded territories.