Add `subscription` to Channel
`channel.subscriptions` returns the list of "Subscription objects"
of the authorized account to the channel. However, there can be at
most one (the account is or is not subscribed to the channel).

For clarity, `has_many :subscriptions` has therefore been replaced
with `has_one :subscription`. The old `subscriptions` method is
still accessible, since that's included in the behavior of `has_one`.