Pass statistics and content details to Video#new
If a user specifies .where(part: 'statistics') when listing videos,
then YouTube will return the "statistics" part and the Video object
will be initialized with that data. Same for content details.

This overrides the default behavior of including status and snippet,
so if those parts are still required, the request should be:

.where(part: 'statistics,snippet,status')