Finished cleaning up Request class
Finally all the methods of Request are documented and agnostic of
YouTube (except :api_key and :camelize_params).

This commit also renames the public method `refresh` on Account to
`refreshed_access_token?`, but this method was never really exposed
as part of the public API, so no major version bump is required.