Update collections/videos.rb default order
Previously, We were to ordering the search results for a query to be 'date' by default. (as in, order by most recently added videos). Change it to be order: 'relevance' instead. That is the what YouTube sorts by by default, and that is what most people expect to get when they do a search using the Yt gem. If they want newest they should explicitly ask for it.