Temporarily return less playlists in reports
Although YouTube Analytics is able to provide 200 playlists at the
time for reports (e.g., views per playlist), returning more than 50
might break the related requests that Yt makes to obtain parts of
each playlist (id, snippet).

This is similar to 02c9405c and could be fixed by keeping the limit
to 200 and by modifying http://git.io/vt38K#L284L294 to fetch the
parts of the playlists in batches of 50.

This commit _partially_ solves the problem by limiting the result to
50 playlists instead.