Add .includes(:video) to .playlist_items
Closes #213

Note that there is currently no way to specify *which* parts of the video
to eager-load: they are hard-coded to snippet, statistics, status.

These are the most common parts, and therefore the ones loaded with the
short syntax `playlist.playlist_items.includes(:video)`.

In the future, we might have `includes` accept a hash (rather than a symbol)
to specify the parts, such as `playlist.playlist_items.includes(video: :status)`.