Fix: print the right error on unauthorized client
Normally, when a request to YouTube API fails, YouTube responds with
a JSON body that includes a Hash inside the "error" key.

However, if your request is an authentication request to get a refresh
token, and the credentials are wrong, then YouTube just responds with:

```json
{"error" : "unauthorized_client"}
```

In this case, the content of the `"error"` key is not a Hash but a
String, so we shouldn’t try to parse a Hash.