Fix Video#update with publishAt
If you look at this section in the documentation, publishAt works as follows:

If you set this property's value when calling the videos.update method, you must also set the status.privacyStatus property value to private even if the video is already private.
If your request schedules a video to be published at some time in the past, the video will be published right away. As such, the effect of setting the status.publishAt property to a past date and time is the same as of changing the video's privacyStatus from private to public.