Better deal with channels with > 500 videos
YouTube has changed the behavior of the API for channels with > 500 videos.
Rather than **not returing** a `nextPageToken`, it now returns one, but it's
a _broken_ one, with duplicate or missing results.

The new approach of YT is to simply switch to the `publishedBefore` way as
soon as 500 videos have been retrieved, since this is what YouTube documents.

This is still not perfect, since each page is cached by YouTube at different
times, and therefore some duplicates occur in successive pages. But it's better
than before, and all we can do without rebuilding YouTube.