Don’t eager load more than 50 assets at once
Calling `content_owner.videos.inclue(:claim)` will result in a first query that
loads 50 videos of the content owner, then a second query that will load all
the claims for those videos, then a third query to fetch the asset of each claim.

The problem is that there can be *more than one claim on each video* by the
same content owner.

For instance, we can fetch 50 videos, then when we fetch claims we get 51.
At that point, if we pass the 51 claim IDs to fetch the assets, we get an error:

```
{"class"=>"RequestError", "errors"=>[{"domain"=>"global", "reason"=>"invalid", "message"=>"The number of asset IDs you provided is too large. You may only list up to 50 assets per request.", "locationType"=>"parameter", "location"=>"id"}], "code"=>400, "message"=>"The number of asset IDs you provided is too large. You may only list up to 50 assets per request."}
```

To fix this, we assume that a user who is eager-loading video > claim > asset
will be happy just loading *one* claim per video. If that's not the case, then
the user should not use eager-loading.