Don’t combine forContentOwner and publishedBefore
This is a new issue that was submitted to YouTube a while ago and has
not been fixed since. Combining the two parameters in the same requests
causes YouTube API to respond with a 400 error that is not documented.

To avoid this, let's just fetch all the videos from Search#list and hope
that YouTube does not return duplicates (as it used to do before).