Re-enable playlist specs
Reenable playlist specs with VCR recordings. Some of the recordings are
a bit fudged in that I added some `sleep` calls since it seems like
playlist-related operations are eventually consistent. This seems OK for
now since the main intent for these tests from now on is to serve to
prevent regressions.

I also removed a test that I couldn't get working that was testing some
quirk with the YouTube API.