:vhs: Record the test suite with VCR (#376)
Changes the test setup to use VCR on any tests that hit the YouTube API.
Also simplifies the use of authentication credentials, and enables (or
deletes) all tests with the exception of those requiring authentication
on behalf of a content owner. Finally, changes `Request` and `Reports`
to make their retry times stubbable in order to speed up test runs.