Don't retry on quota errors (#401)
* Don't retry on quota errors

They say that insanity is doing the same thing over and over and
expecting different results...

This patch removes retry behavior on quota issues because:

* this isn't useful when daily quota has been reached
* in the case of 100s quota, you'd need to be attempting requests ~61s
into the quota cool-off in order for the final retry to succeed. That
seems pretty unlikely.

So for all practical purposes, when quota is reached, the library stalls
for the maximum timeout time (39s by my math). This is further
problematic when making requests in environments like Heroku that have
a 30s timeout for web request-response cycles.

Let's do the sane thing and just raise an error when quota is reached.

* Bump to 0.33.4 [ci-skip]